fix: switch people models not in correct file name
